Merge consecutive rasterizations
In vector output it is possible to flag artists to be rasterized. In
many cases with multiple rasterized objects there can be significant
file size savings by combining the rendered bitmaps into a single
bitmap.

This is achieved by moving the depth tracking logic from
start_rasterizing() and stop_rasterizing() functions into the
allow_rasterization() wrapper. This allows delaying the call to
stop_rasterizing() until we are about to draw an non rasterized artist.
stop_rasterizing() must also be called from the finalize() method.

This fixes #17149